backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pillow==10.4.0
numpy==2.1.3
tensorflow==2.20.0
pymongo==4.13.2
